requirements.txt
# Core scientific computing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning and clustering
scikit-learn>=1.0.0
umap-learn>=0.5.5

# Visualization
matplotlib>=3.4.0

# Data handling
bitarray>=2.3.0
requests>=2.26.0